[Refactor] Separate Weekday
