+++ Chinggu/Chinggu/Utils/Weekday.swift
@@ -1,0 +1,30 @@
+//
+//  Weekday.swift
+//  Chinggu
+//
+//  Created by Sebin Kwon on 2023/08/30.
+//
+
+import Foundation
+
+enum Weekday: String, CaseIterable {
+    case monday = "월요일"
+    case tuesday = "화요일"
+    case wednesday = "수요일"
+    case thursday = "목요일"
+    case friday = "금요일"
+    case saturday = "토요일"
+    case sunday = "일요일"
+    
+    var weekdayValue: Int {
+        switch self {
+        case .sunday: return 1
+        case .monday: return 2
+        case .tuesday: return 3
+        case .wednesday: return 4
+        case .thursday: return 5
+        case .friday: return 6
+        case .saturday: return 7
+        }
+    }
+}
